Attemptative to draw the network graph
Added gui module to draw the network graph.
The problem is the bad position of nodes.
It would be nice to drag the node on  the desired position.
Matlibplot doesn't seem to bet the best library to work with graph.
Consider transfer the graph in JSON and use javascirpt
library.